Major Updates
Have updated the particle effects. Updated Upgrade system and UI. updated bugs and fixed many errors.

+++ ElectriciTD_v01/Assets/Scripts/Player/TurretBlueprints.cs
@@ -16,7 +16,7 @@ public class TurretBlueprints : ScriptableObject
 	public float fireRate;
 	public float turnSpeed;
 
-	public int attackValue;
+	public float attackValue;
 	public int health;
 	public int energyCost;
 	public int sellPrice;
